Added StringResponse.class
This is the first of the two classes that will replace SmsOperator,
maintaining functions and improving flexibility.
The added class, StringResponse, will be useful to search a pattern into
a string (an SMS message in our case) and trigger an automated response,
that is to say another set predefined strings related to the pattern.